+++ rvm/src/vm/arch/powerPC/compilers/baseline/GCMap/VM_BaselineGCMapIterator.java
@@ -106,7 +106,7 @@ public final class VM_BaselineGCMapIterator extends VM_GCMapIterator
       VM_Address        ip                       = VM_Magic.getNextInstructionAddress(fp);
       int               callingCompiledMethodId  = VM_Magic.getCompiledMethodID(fp);
       VM_CompiledMethod callingCompiledMethod    = VM_CompiledMethods.getCompiledMethod(callingCompiledMethodId);
-      int               callingInstructionOffset = ip.diff(VM_Magic.objectAsAddress(callingCompiledMethod.getInstructions()));
+      int               callingInstructionOffset = ip.diff(VM_Magic.objectAsAddress(callingCompiledMethod.getInstructions())).toInt();
 
       callingCompiledMethod.getDynamicLink(dynamicLink, callingInstructionOffset);
       bridgeTarget                = dynamicLink.methodRef();
